requirements.txt
telethon
google-genai
openai
pysocks
httpx[socks]
pandas
openpyxl
cryptg
fastapi
uvicorn[standard]
jinja2
python-multipart
itsdangerous
psycopg2-binary
python-telegram-bot
